Work around Trusty CI bindep issue
Ubuntu Trusty is not a supported platform for Ocata onwards.
Currently the CI system is executing jobs against Trusty
unintentionally and it's taking some time to unwind it.

This bypasses the attempt to install a package that isn't
available for Trusty.

Change-Id: I25f0cd4860c1d53dc364eacfd8f397660b6efe05
(cherry picked from commit d7141eec5261930ad4991307e1893e800e2ab75c)