Reduce number of locust clients
When measuring availability, multiple clients effectively give you
additional resolution, but that resolution may be of negligible value
here.

More importantly, when measuring for response time consistency
(important for zero-impact upgrades), it's counter-productive to run
clients in parallel, as you're also measuring the effect of parallel
clients instead of just the impact of the upgrade process.

Change-Id: Ic9449a1050651a1285bf80e69c958c965c21cbc4