Fix rolling upgrade test
The rolling upgrade test is currently not setting
the endpoint correctly due to 'keystone_upgrade'
not being defined.

This patch removes the requirement to define the
var by ensuring that the role always uses haproxy.
This prevents having to remember to set the var
when doing development and makes better sense for
test purposes anyway.

This patch also serialises the upgrade and ensures
that the backend is set into maintenance mode when
the upgrade is actioned.

When a playbook runs os_keystone in serial, the
SSH and fernet key distribution are broken. This
fixes both items allowing the role to be run in a
serialized playbook.

Combined backport of:
https://review.openstack.org/438219
https://review.openstack.org/475040

Change-Id: I8f16495607abb871390d28c0b3e9b2b856dda097