Rebuild credential-key repo during keystone[0] rebuild
When the first Keystone container is rebult in an existing environment,
the credential key repository is overwritten with new keys and the
existing keys are overwritten on the other infrastructure hosts without
any migration taking place. This results in an irrevocable loss of
the keys used to encrypt the credentials.

Now we will collect keys from any existing credential keys on the other
containers and use them to rebuild the credential-key repo on the primary
container before performing a key migration and rotation.

If no keys are found on the other containers, we will perform a
credential_setup on the primary container and sync the keys, just
as we would have before.

Closes-Bug: #1667960
Change-Id: Ic616d397574573629273838fbf68ea3f6bdb0468