Allow the split of install and config
If keystone install is done separately from the config (filtering
or interruption), a re-run of the role will not mark the venv
as changed, and the db sync will not trigger.

Here we force a flush on file next to the venv change, and use
this file (local fact) as a way to know if a db sync is required.

When a dbsync is considered done, i.e after the expand and migrate
stages, we mark the db_migrate as done for all the nodes.

Change-Id: I318700582a65cc1ea6571d9460fd58bed949a4e9
Depends-On: Ie4a47e85bf31e1ac3a2bd6f0dec510bb24abdbbd