Use a custom git checkout in developer mode
This patch allows a developer to customise the code for
keystone, then test the changes when executing any of
the tox tests we implement.

The workflow means that by default the git_repo will be cloned from
upstream (this ensures gating continues to work). If the developer
specifies a "keystone_dev_local_directory" then the repo will not be
cloned, and it assumes the developer has already cloned and edited that
repo. A clone can be forced using "keystone_dev_force_clone" to "True".

Change-Id: Ibb666a803e73b50e3ee1918e633a9bbb84b084da