Use uri module for git sourced configs
The current mechanism uses a lookup, causing the
fetch of the default templates to happen via curl
from the deployment node. This causes problems if
the deployment node does not have access to the
repo server web service, which may be the case in
high security environments.

This patch changes the mechanism to only use the
lookup module for the file content that is on the
deployment node, then falls back to using the uri
module to fetch the default content. This ensures
that the deployment node does not have to reach
into the environment for the content.

Also, in https://review.openstack.org/446235 a
new task was added to template the paste/policy
files from the git source but the owner/group/mode
settings were not implemented. This patch
consolidates the two config_template tasks into
one and ensures that the owner/group/mode settings
are properly implemented for all the files.

Change-Id: I47de2e4b1c864035c8b686e96e65722f312b7280