Use command instead of debug for handlers
Due to the debug message plugin the handler restart
messages show at the end of the playbook execution
which is a little confusing. Using debug also
requires setting changed_when to true which is a
little extra bit of code which we do not have to
carry.

Instead we use the command module which is simple,
works and less wordy.

Change-Id: I9e3acb966c0d16b61fbc998642e36f4015064155
(cherry picked from commit 64a01c573b945245e1f56f58b39c836f9074815c)