Convert old fact format to the new format
In [1] the fact format was changed, but no
migration from the old format to the new format
was implemented.

Deployments therefore implemented using the old
format will fail when upgrading to the next tag
due to the appropriate fact being missing.

This patch bridges the gap to ensure that minor
upgrades work as expected.

[1] https://review.openstack.org/#/q/Icf443a93424c9eba7b68d3279ebb48ae00602808

Change-Id: I9becdf59a36c2a1255ee51728dd764ef5dc79790