Fix rolling upgrade test
The rolling upgrade test is currently not setting
the endpoint correctly due to 'keystone_upgrade'
not being defined.

This patch removes the requirement to define the
var by ensuring that the role always uses haproxy.
This prevents having to remember to set the var
when doing development and makes better sense for
test purposes anyway.

This patch also serialises the upgrade and ensures
that the backend is set into maintenance mode when
the upgrade is actioned.

Change-Id: I8f16495607abb871390d28c0b3e9b2b856dda097
Depends-On: I5cbb3824430dc09b36476f81e0cdfd4f0a15f497