Split user create and ssh key generation
In order to allow an install and config split, but not
to have ssh keys left inside an pre-installed container,
the two tasks are split and tagged appropriately.

Backport note:
ansible_play_hosts is not used for this backport due
to the different way in which Ansible 2.1 implements
it.

Change-Id: I468d1178179d70edfe4b19d40a9a32b35ad18258
(cherry picked from commit 43d03a03cca7bdf81cad0936cdd6f479b1d70eb7)