Correct behaviour for serialised execution
Ansible 2.1 does not use ansible_play_hosts in
the same was as Ansible 2.2 - it provides the
hosts in batches instead of the whole set.

This patch corrects the behaviour to cater for
that.

Due to the recent change in the tests repo which
added haproxy backend disabling in the keystone
playbook, empty host groups for haproxy are added
to the inventory. This ensures that the playbook
can execute without errors when there is more
than one keystone container. A follow-up patch
will enable the use of haproxy for the role tests.

Depends-On: I2f6e1ff4526a60a581f01c2381cec43332a5560b
Change-Id: Ieb04c2389954442762af70c66e4d204483a64bfd