Implement serialisable rolling upgrade
The current upgrade implementation executes a db
sync while all keystone services are online
while the older software is still running,
potentially resulting in data corruption.

This patch ensures that the role is serialisable
and that it executes the rolling upgrade
mechanism available in newton.

This is a re-implementation of several patches
which have been implemented in ocata/pike.

Change-Id: I5650f16b9a115bd392012b743788057a94d09226