Replace db sync check with alternative process
The keystone db sync --check option is only available
in Pike. Trying to use this option causes the db sync
processes to run every time the playbook is executed,
resulting in the keystone service being stopped. The
service is also never notified to restart, so it never
does.

This patch uses just the local facts to determine
whether or not to execute the db sync actions, and
properly notifies the service to restart afterwards.

Note:
This includes a re-implementation of https://review.openstack.org/483309
which is different due to the fact that master exclusively uses uwsgi.

Change-Id: Ide64927e43e7684f03be7a73b893283c80e89afc