Ensure online migrations can run
The online migrations are disabled when they would have run
once per software version. However, when the task that
sets data_migration is skipped, it's not considered as failure
by ansible, and therefore the | succeeded will be true, removing
the ability to do the online data migrations.

This ensure that the fact task will be set to remove the need
for online data migrations when the relevant task has properly
run (not skipped).

Change-Id: I2a20974fbfdf20ce9675a91599f3d654cad03ffc
Closes-Bug: #1713698