Use the tests ansible-previous venv
When doing the role upgrade, we run every playbook with current's
branch ansible. This is normal for the infrastructure bits (which
are running on current branch), but not for the installation of
the previous branch's role, which should use previous branch's
environment.

The tests repo now has this venv, so we should make use of it.

Depends-On: I7ebb045885dd645b820de2b7f75b46c755c015f6
Change-Id: I579a3f7b641e02f40ed1b21f523aecbb9f16835a
Related-Bug: #1741471