Use the good configuration directive for memcache servers
Under Pike, both the backend argument and memcache_servers work.
They both end-up with connections to the backend of memcache.

The latter doesn't have the bug when multiple servers are given,
though. So let's use this one exclusively.

This massive template change is for backwards compatibility:
If a deployer has set on of those keystone variables, the servers
used won't change.

Change-Id: I7e29aa0a9055b737dd4cf171c40e2105e4b32e94
Related-Bug: #1743036