Add ability to configure NGINX to show correct IP in logs
At the moment, NGINX will be using the IP which initiated the
connection in the logs which could be misleading if you are using
it behind a load balancer.

This patch adds the ability to be able to configure the hosts
which can send the semi-standard X-Forwarded-For header in
order to show the right IP in the logs.

Change-Id: I9006dbb15837178cee7de869372c901040095e2b