Only replace python when re-initializing the venv
We reinitialize the venv to ensure that the right version of
python is in the venv, but we do not want virtualenv to also
replace pip, setuptools and wheel so we tell it not to. If we
don't do this then virtualenv will install the latest available
version, which is not what we want.

Change-Id: I0f0751a11dffbd77eaf2bfd238f2f934f11847c8
Partial-Bug: #1764470
(cherry picked from commit 4d92214d6b0ae93f0fd28d67783d7bf25d5d3234)