Do not pip install keystoneclient on the host
The keystoneclient package is being installed on the host by PIP but
that means that a whole bunch of required dependencies are being pulled
in as well.

This brings the host to a rather messed up state when installing
keystone from distro packages, since distribution and
PIP packages are being mixed together. We can simply avoid that by
only using the virtualenv where required, in which the keystone client
would be installed. We still require the virtualenv package to be
installed on the host from PIP at the moment to get the expected
recent version of pip inside the virtualenv.

Depends-On: I5a78e2120e596d36629b4ba978b2b5df76b149b0
Depends-On: Ib64dcbc960df7d369d202ce8cf7bdc29b3ee0e0a
Change-Id: Ia51a1f4081b3a56819237d0352cc27c1f32f433c
Implements: blueprint openstack-distribution-packages