Use a nginx conf.d file for the custom logging
To cater for the situation where there are no containers, or
nginx is shared in some way, it's better to add custom directives
for nginx.conf into a conf.d file. That way it's never overwritten.

To resolve the situation where repo_server overwrites the config
laid down by os_keystone, we implement the custom log in this way.

The task which verifies that the right config directories exist
is modified to ensure that both the sites and the conf.d directories
exist before this configuration is laid down. For CentOS/OpenSUSE
they are the same directory, but for Ubuntu they are different.

This patch also ensurs that the previous implementation is removed,
otherwise an upgrade fails. This patch should be back portable to
ensure that zero-container deploys for Queens work properly.

Change-Id: I2ab0f591c1cdeae43a002d9c8cf9bbfdd69597f1