Ensure __pycache__ folder is deleted
When running python3, the cleanup of
the virtualenv is incomplete: While the *.pyc files are removed,
the python3 __pycache__ folder is not removed.

That causes the sed instruction (applying to files) to fail,
as the __pycache__ folder is not a regular file, but a folder.

With the sed not successfully running, the shebang of the
virtualenv is pointing to an incorrect location, causing then
a failure in keystone-manage commands.

On top of it, a success would happen should the last command
return 0. We now make sure the whole script must return
success.

This should fix the problem.

Change-Id: Id28fd1c4989bfff1d0d736c037b8f0e77eb082c1