tasks: Record a venv_tag local fact for distro installations
Distro installations do not deploy virtual environments but they still
need a local venv_tag deployed. The reason for this is because this fact
is used by the integration repo to determine if the keystone software is
the same across all nodes in the keystone_all group so we can safely run
the DB migration.

Change-Id: I9bab11b3c643f2411d78267aa6114081fe76b9ef