Import GPG key for NGINX via Ansible
The GPG key for NGINX was added manually therefore YUM would try
to pull it down and fail because the repository cache would not
allow any requests that tried to pull down a GPG key.

This patch adds it via Ansible beforehand that way it doesn't go
through the proxy.

Change-Id: I96ad7f049ab72ce338596ae34362e3a4ca026bb6