Upgrade pip/setuptools/wheel in role test venv
In the integrated build, the repo build process upgrades
pip/setuptools/wheel in the venv to our preferred pinned
version, but developer mode does not.

Keystone requires the pysaml2 python library, which (from
4.6.0) due to [1] requires setuptools > 30.3.0 - but CentOS
only has 28.8.0 when it builds the venv, so the package
install fails.

In this patch we create the venv, then upgrade pip,
setuptools and wheel, then install the keystone packages.

Doing this for all roles is impractical, but it will be
necessary to unblock patches from merging into roles today.
In the future when we use the python_venv_build role we can
do all this there instead of repeating this patch across all
the roles.

[1] https://github.com/IdentityPython/pysaml2/commit/9c7b0014ccc4229031813ddcc5f6790b59540df5
Change-Id: I96434ccc7eb5f02e12605e378d733c39f092a411