Define cache_timeout variable
Much like the 'debug' variable, the 'cache_timeout' variable
is an expected variable, but is currently not defined in the
role. This causes the role to fail when the galera_client
meta-dep is removed (because that role defined it, this role
inherited it).

Change-Id: I97ba0912ff9ef065cc0184ebc4ee8cd68d919c47
(cherry picked from commit 484fb899a0b3b12fc05865e8678664fc20679d68)